business-intelligence-service/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
confection==0.1.5
cymem==2.0.11
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
networkx==3.5
nltk==3.9.1
numpy==2.3.0
packaging==25.0
preshed==3.0.10
pydantic==2.11.6
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
thinc==8.3.6
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
